Scape_V2/demo/requirements.txt
flask>=3.1.1
werkzeug>=2.0.0
PyPDF2>=3.0.1
requests>=2.32.2
python-dotenv>=0.19.0